requirements.txt
Django>=5.2
djangorestframework>=3.14
djangorestframework-simplejwt>=5.4.0
